core: move react app to where frontend expects it
